requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
requests
google-generativeai
python-dotenv
pydantic-settings
pytest
httpx
